Implemented app router.
