+++ lib/db/mongoose.js
@@ -14,7 +14,9 @@ module.exports = function(dbConfig, cb){
 		cb(db, new mongoSession({
 			collection: 'fw.sessions',
 			mongoose_connection: dbConnection
-		}));
+		}), function(cb){
+			db.disconnect(cb);
+		});
 	});
 	dbConnection.on('error', function(err){
 		console.error(err);
