+++ lib/default/fwconfig.js
@@ -10,6 +10,7 @@ if(envConfig.charAt(0) !== '{') {
 
 module.exports = {
 	mode: (envConfig.mode || 'DEFAULT').toLowerCase(), // the running mode
+	app: [], // the app(s) need to be initially started
 	ip: envConfig.ip || '', // the ip address to bind, default binded to any address
 	port: envConfig.port || [80], // the port(s) to bind (it may requires root privilege to bind to 80, and it's dangerous!)
 	websocket: envConfig.websocket || false, // whether websocket should be enabled (some cloud platforms does not support websocket)
