+++ lib/main.js
@@ -3,230 +3,122 @@
 
 var fs = require('fs');
 var http = require('http');
-var express = require('express');
-var session = require('express-session');
-var cookieSignature = require('express-session/node_modules/cookie-signature');
-var compress = require('compression');
-var cookieParser = require('cookie-parser');
-var sockjs = require('sockjs');
-var locale = require('locale');
 
-var tmplParser = require('./tmpl_parser.js').parser;
-var fsListener = require('./fs_listener.js');
-var ipsParser = require('./ips_parser.js');
-var utils = require('./utils.js');
 var appCreator = require('./app_creator.js');
+var tmplParser = require('./tmpl_parser.js').parser;
 
 // create global fw object
-global.fw = {};
+var fw = global.fw = {};
+
+// add fw methods
+var fwObjInit = function(){
+	// basic properties
+	fw.config = JSON.parse(process.env.FW);
+	fw.debug = (fw.config.mode === 'debug');
+	fw.restart = function(){
+		if(fw.mode !== 'debug' && fw.mode !== 'cache' && fw.mode !== 'run') throw('Cannot restart in current running mode (' + fw.mode + ').');
+		else process.exit(250);
+	};
+	// loaders
+	fw.currentLoading = null;
+	fw.tmpl = function(file){
+		if(!fw.currentLoading) return null;
+		file = fw.currentLoading.slice(0, fw.currentLoading.lastIndexOf('/')+1) + file;
+		return tmplParser(app, fs.readFileSync(file).toString('utf8'), file);
+	};
+	fw.module = function(name){
+		// TODO module path
+		if(typeof(moduleRes[name]) !== 'undefined')
+			return moduleRes[name];
+		if(name.slice(-3) !== '.js')
+			return moduleRes[name+'.js'];
+	};
+	// app loader
+	fw.createApp = function(path){
+		var args = [appRouter];
+		for(var i=0; i<arguments.length; i++) {
+			args.push(arguments[i]);
+		}
+		var app = appCreator.apply(fw, args);
+		return app;
+	};
+};
+
+// app manager and router
+var appRouter = (function(){
+	var hostRoute = {};
+	var bindHost = function(app, host){
+		host = host || '';
+		if(hostRoute[host]) hostRoute[host].push(app);
+		else hostRoute[host] = [app];
+	};
+	var unbindHost = function(app, host){
+		host = host || '';
+		if(!hostRoute[host]) return;
+		for(var i=0; i<hostRoute[host].length; i++) {
+			if(hostRoute[host][i] === app) {
+				hostRoute[host].splice(i, 1);
+				i--;
+			}
+		}
+		if(!hostRoute[host].length) delete hostRoute[host];
+	};
+	var routeReq = function(req){
+		var arr = hostRoute[req.headers.host] || hostRoute[''];
+		if(arr) return arr[0];
+	};
+	return {
+		bindHost: bindHost,
+		unbindHost: unbindHost,
+		routeReq: routeReq
+	};
+})();
 
 // main
 var main = function(config){
-	// parse extra params
-	var initialApps = [];
-	for(var i=1; i < arguments.length; i++) {
-		initialApps = initialApps.concat(arguments[i]);
-	}
-
-	// init fw object
-	fw.restart = function(){
-		if(fw.mode !== 'debug' && fw.mode !== 'cache' && fw.mode !== 'run') throw('Cannot restart in current mode (' + fw.mode + ').');
-		else process.exit(250);
-	};
-
-	// init express
-	var app = express();
-	app.use(compress());
-	app.use(cookieParser());
-
-	// config
-	// TODO
-	fw.config = utils.deepExtend(config);
-
-	// get the selected language, hacked into locale
-	locale.Locale['default'] = '';
-	var supportedLocale = new locale.Locales(app.config.app.locale);
-	app.use(function(req, res, next){
-		req.selectLanguage = function(str){
-			req.language = (new locale.Locales(str)).best(supportedLocale).toString();
-			if(typeof(str) !== 'undefined')
-				res.cookie('fw.lang', req.language, {expires: new Date(Date.now() + 86400*365)});
-			else
-				res.clearCookie('fw.lang');
-		};
-		if(typeof(req.cookies['fw.lang']) !== 'undefined') var lang = req.cookies['fw.lang'];
-		else var lang = req.headers['accept-language'] || '';
-		req.language = (new locale.Locales(lang)).best(supportedLocale).toString();
-		next();
-	});
-
-	// conn object builder
-	app.connRpc = function(conn){
-		return function(path){
-			// parse args
-			var args = [];
-			for(var i=1; i<arguments.length; i++)
-				if(typeof(arguments[i]) === 'function') {
-					var doneCb = arguments[i];
-					if(typeof(arguments[i+1]) === 'function')
-						var errCb = arguments[i+1];
-					break;
-				} else {
-					args.push(arguments[i]);
-				}
-			// routing
-			var funcs = app.routeMethod(path);
-			if(!funcs.length) return;
-			// generate res func
-			var cb = function(){
-				if(!doneCb) return;
-				var a = [];
-				for(var i=0; i<arguments.length; i++)
-					a.push(arguments[i]);
-				doneCb.apply(cb, a);
-			};
-			cb.err = function(){
-				if(!errCb) return;
-				var a = [];
-				for(var i=0; i<arguments.length; i++)
-					a.push(arguments[i]);
-				errCb.apply(cb, a);
-			};
-			cb.next = function(){
-				var func = funcs.shift();
-				if(func) func.apply(conn, args);
-			};
-			args.unshift(conn, cb);
-			cb.next();
-		};
-	};
-	app.use(function(req, res, next){
-		if(app.debug) console.log(nowString()+' '+req.method+': '+req.path);
-		next();
-	});
+	fwObjInit();
 
 	// init sockjs
-	app.sock = sockjs.createServer({
-		heartbeat_delay: app.config.socket.heartbeat,
-		websocket: !app.config.server.disableWebsocket,
+	var sockServer = sockjs.createServer({
+		heartbeat_delay: fw.config.heartbeat,
+		websocket: !fw.config.websocket,
 		log: function(type, msg){
-			if(app.debug && type !== 'debug')
-				console.log(nowString()+' Socket: '+msg);
+			if(fw.debug && type !== 'debug' && type !== 'info')
+				console.log(utils.nowString() + ' Socket ' + type + ': ' + msg);
 		}
 	});
-	var ports = app.config.server.port;
-	if(ports.constructor !== Array) ports = [ports];
-	var servers = [];
-	for(var i=0; i<ports.length; i++) {
-		var server = http.createServer(app);
-		app.sock.installHandlers(server, {prefix:'/~sock'});
-		servers.push(server);
-	}
-
-	// session handler
-	var sessionStore = null;
-	var sessionProxy = function(options){
-		if(!options.store)
-			options.store = new session.MemoryStore();
-		sessionStore = options.store;
-		return session(options);
-	};
-	app.getAuthSession = function(auth, cb){
-		if(auth.charAt(1) !== ':') {
-			cb(null);
-			return;
+	sockServer.on('connection', function(sock){
+		sock.app = appRouter.routeReq(sock);
+		if(sock.app) {
+			if(fw.debug) console.log(utils.nowString() + ' Connected: ' + sock.protocol);
+			req.app.socket(sock);
+		} else {
+			sock.end();
 		}
-		var str = cookieSignature.unsign(auth.slice(2), app.config.secret.cookie);
-		sessionStore.get(str, function(error, res){
-			if(error || !res) {
-				cb(null);
-				return;
+	});
+
+	// create node http server
+	var ports = fw.config.port;
+	if(ports.constructor !== Array) ports = [ports];
+	for(var i=0; i<ports.length; i++) {
+		var server = http.createServer(function(req, res){
+			// express proxy for different host
+			req.app = appRouter.routeReq(req);
+			if(req.app) {
+				req.app.express(req, res);
+			} else {
+				res.writeHead(503);
+				res.end();
 			}
-			var sess = new session.Session({
-				sessionID: str,
-				sessionStore: sessionStore
-			}, res);
-			cb(sess);
 		});
-	};
-	app.generateSession = function(cb){
-		var req = {};
-		sessionStore.generate(req);
-		sessionStore.set(req.sessionID, req.session, function(){
-			var s = 's:' + cookieSignature.sign(req.sessionID, app.config.secret.cookie);
-			cb(s);
-		});
-	};
-	app.sessionCookieVal = function(id){
-		return 's:' + cookieSignature.sign(id, app.config.secret.cookie);
-	};
+		sockServer.installHandlers(server, {prefix: '/~sock'});
+		server.listen(ports[i], fw.config.ip);
+	}
+	console.log('fw.mpa started at port(s) ' + ports.join(', ') + '.');
 
-	// connect db
-	var initDb = function(cb){
-		if(app.debug) console.log('Connecting to database: ' + app.config.db.type);
-		var dbEngine = require('./db/' + app.config.db.type + '.js');
-		dbEngine(app.config.db, function(db, sessionStore){
-			app.db = db;
-			app.use(sessionProxy({
-				key: 'fw.sid',
-				secret: app.config.secret.cookie,
-				cookie: {
-					maxAge: app.config.server.sessionLifeTime
-				},
-				store: sessionStore,
-				resave: true,
-				saveUninitialized: true
-			}));
-			setTimeout(cb, 0);
-		});
-	};
-
-	// route and listen
-	var route = function(){
-		require('./http_router.js').route(app, express, function(){
-			require('./sock_router.js').route(app, function(){
-				for(var i=0; i<ports.length; i++) {
-					servers[i].listen(ports[i], app.config.server.ip);
-				}
-				console.log('Server started at port(s) ' + ports.join(', ') + '.');
-			});
-		});
-	};
-
+	// TODO
 	// start server
 	var startServer = function(){
-		// generate app object for server codes
-		fw = app.obj = {
-			debug: app.debug,
-			config: app.config,
-			db: app.db,
-			restart: app.restart,
-			tmpl: function(file){
-				if(!fw.currentLoading) return null;
-				file = fw.currentLoading.slice(0, fw.currentLoading.lastIndexOf('/')+1) + file;
-				return tmplParser(app, fs.readFileSync(file).toString('utf8'), file);
-			},
-			module: function(name){
-				if(typeof(moduleRes[name]) !== 'undefined')
-					return moduleRes[name];
-				if(name.slice(-3) !== '.js')
-					return moduleRes[name+'.js'];
-			}
-		};
-		// req.conn middleware
-		app.use(function(req, res, next){
-			req.conn = {
-				session: req.session,
-				host: req.get('Host'),
-				ips: ipsParser(req.header('X-Forwarded-For')),
-				ip: req.ip,
-				headers: req.headers,
-				language: req.language,
-				selectLanguage: req.selectLanguage,
-			};
-			req.conn.rpc = app.connRpc(req.conn);
-			next();
-		});
 		// loading server modules
 		var moduleRes = {};
 		if(fs.existsSync(app.config.path.module)) {
@@ -260,8 +152,6 @@ var main = function(config){
 			route();
 		}
 	};
-
-	initDb(startServer);
 };
 
-module.exports = main;
+main();
