+++ lib/routers/static.js
@@ -1,0 +1,4 @@
+// Copyright 2014 LastLeaf, LICENSE: github.lastleaf.me/MIT
+'use strict';
+
+module.exports = function(app, prefix, path, cb){};
