+++ lib/utils.js
@@ -4,7 +4,7 @@
 // deep extend
 exports.deepExtend = function(dest, src){
 	for (var prop in src) {
-		if (typeof src[prop] === "object" && src[prop] !== null ) {
+		if (typeof src[prop] === 'object' && src[prop] !== null) {
 			dest[prop] = dest[prop] || {};
 			deepExtend(dest[prop], src[prop]);
 		} else {
